--- src/main/java/com/oleynik/qa/workshop/Factorial.java
@@ -1,9 +1,0 @@
-package com.oleynik.qa.workshop;
-
-public class Factorial {
-    public static long factorial(long number){
-        if (number <= 1)
-            return 1;
-        else return (number * factorial(number - 1));
-    }
-}
