--- src/main/java/com/oleynik/qa/workshop/Utils.java
@@ -1,11 +1,0 @@
-package com.oleynik.qa.workshop;
-
-public class Utils {
-    public static void waitFor(int timeoutInSeconds) {
-        try {
-            Thread.sleep(timeoutInSeconds * 1000L);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-}
