+++ src/main/java/com/oleynik/qa/workshop/junit/Factorial.java
@@ -1,0 +1,9 @@
+package com.oleynik.qa.workshop.junit;
+
+public class Factorial {
+    public static long factorial(long number){
+        if (number <= 1)
+            return 1;
+        else return (number * factorial(number - 1));
+    }
+}
