+++ src/main/java/com/oleynik/qa/workshop/junit/Utils.java
@@ -1,0 +1,11 @@
+package com.oleynik.qa.workshop.junit;
+
+public class Utils {
+    public static void waitFor(int timeoutInSeconds) {
+        try {
+            Thread.sleep(timeoutInSeconds * 1000L);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
