--- src/test/java/com/oleynik/qa/workshop/ddt/DataProvider2Test.java
@@ -1,19 +1,0 @@
-package com.oleynik.qa.workshop.ddt;
-
-import com.oleynik.qa.workshop.Factorial;
-import com.oleynik.qa.workshop.junit.dataproviders.MyDataProvider;
-import com.tngtech.java.junit.dataprovider.DataProviderRunner;
-import com.tngtech.java.junit.dataprovider.UseDataProvider;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-@RunWith(DataProviderRunner.class)
-public class DataProvider2Test {
-
-    @Test
-    @UseDataProvider(value = "read_numbers", location = MyDataProvider.class)
-    public void junit4_with_external_data_provider_test(long number, long expected) {
-        Assert.assertEquals("Factorial function is wrong.", expected, Factorial.factorial(number));
-    }
-}
