--- src/test/java/com/oleynik/qa/workshop/grouping/categories/SmokeTests.java
@@ -1,4 +1,0 @@
-package com.oleynik.qa.workshop.grouping.categories;
-
-public interface SmokeTests {
-}
