+++ src/test/java/com/oleynik/qa/workshop/junit/ddt/Parameterization2Test.java
@@ -1,0 +1,30 @@
+package com.oleynik.qa.workshop.junit.ddt;
+
+import com.oleynik.qa.workshop.junit.Factorial;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameter;
+import org.junit.runners.Parameterized.Parameters;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+@RunWith(Parameterized.class)
+public class Parameterization2Test {
+    @Parameter
+    public long number;
+    @Parameter(value = 1)
+    public long expected;
+
+    @Parameters
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{{1, 1}, {2, 2}, {3, 6}, {4, 24}, {5, 120}});
+    }
+
+    @Test
+    public void junit4_ddt_v2_factorial_test() {
+        Assert.assertEquals("factorial function works wrong", expected, Factorial.factorial(number));
+    }
+}
