+++ src/test/java/com/oleynik/qa/workshop/junit/grouping/categories/RegressionTests.java
@@ -1,0 +1,4 @@
+package com.oleynik.qa.workshop.junit.grouping.categories;
+
+public interface RegressionTests {
+}
