--- src/test/java/com/oleynik/qa/workshop/suites/MySuite.java
@@ -1,11 +1,0 @@
-package com.oleynik.qa.workshop.suites;
-
-import com.oleynik.qa.workshop.general.AssertTest;
-import com.oleynik.qa.workshop.general.MyRunnerTest;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({AssertTest.class, MyRunnerTest.class})
-public class MySuite {
-}
